+++ sqldata/sqlsncapi.h
@@ -28,7 +28,7 @@
 #include <sql.h>
 #include <sqlext.h>
 #include <sqltypes.h> 
-#include <sqlncli.h>
+#include "sqlncli.h"
 #include <oci.h>
 #include <ctpublic.h>
 #include "sqlapibase.h"
@@ -89,14 +89,14 @@ class SqlSncApi : public SqlApiBase
 	int _cursor_fetched;
 
 	// Number of columns in bulk copy operation
-	size_t _bcp_cols_count;
+	long _bcp_cols_count;
 	// Column describtions and data for bulk copy
 	SqlCol *_bcp_cols;
 	// Exist LOB columns to be sent using bcp_moretext
 	bool _bcp_lob_exists;
 
 	// Codepage to use for BCP
-	int _bcp_codepage;
+	long _bcp_codepage;
 
 	// SQL Server Native Client DDL (for Bulk Copy API)
 #if defined(WIN32) || defined(_WIN64)
@@ -157,30 +157,30 @@ public:
 	virtual void Deallocate();
 
 	// Get row count for the specified object
-	virtual int GetRowCount(const char *object, int *count, size_t *time_spent);
+	virtual int GetRowCount(const char *object, long *count, size_t *time_spent);
 
 	// Execute the statement and get scalar result
-	virtual int ExecuteScalar(const char *query, int *result, size_t *time_spent);
+	virtual int ExecuteScalar(const char *query, long *result, size_t *time_spent);
 
 	// Execute the statement
 	virtual int ExecuteNonQuery(const char *query, size_t *time_spent);
 
 	// Open cursor and allocate buffers
-	virtual int OpenCursor(const char *query, size_t buffer_rows, int buffer_memory, size_t *col_count, size_t *allocated_array_rows, 
-		int *rows_fetched, SqlCol **cols, size_t *time_spent, bool catalog_query = false,
+	virtual int OpenCursor(const char *query, long buffer_rows, long buffer_memory, long *col_count, long *allocated_array_rows, 
+		long *rows_fetched, SqlCol **cols, size_t *time_spent, bool catalog_query = false,
         std::list<SqlDataTypeMap> *dtmap = NULL);
 
 	// Fetch next portion of data to allocate buffers
-	virtual int Fetch(int *rows_fetched, size_t *time_spent);
+	virtual int Fetch(long *rows_fetched, size_t *time_spent);
 
 	// Close the cursor and deallocate buffers
 	virtual int CloseCursor();
 
 	// Initialize the bulk copy from one database into another
-	virtual int InitBulkTransfer(const char *table, size_t col_count, size_t allocated_array_rows, SqlCol *s_cols, SqlCol **t_cols);
+	virtual int InitBulkTransfer(const char *table, long col_count, long allocated_array_rows, SqlCol *s_cols, SqlCol **t_cols);
 
 	// Transfer rows between databases
-	virtual int TransferRows(SqlCol *s_cols, int rows_fetched, int *rows_written, size_t *bytes_written,
+	virtual int TransferRows(SqlCol *s_cols, long rows_fetched, long *rows_written, size_t *bytes_written,
 		size_t *time_spent);
 
 	// Specifies whether API allows to parallel reading from this API and write to another API
@@ -195,11 +195,11 @@ public:
 	virtual int DropReferences(const char* table, size_t *time_spent);
 
 	// Get the length of LOB column in the open cursor
-	virtual int GetLobLength(size_t row, size_t column, size_t *length);
+	virtual int GetLobLength(long row, long column, long *length);
 	// Get LOB content
-	virtual int GetLobContent(size_t row, size_t column, void *data, size_t length, int *len_ind);
+	virtual int GetLobContent(long row, long column, void *data, long length, long *len_ind);
 	// Get partial LOB content
-	virtual int GetLobPart(size_t row, size_t column, void *data, size_t length, int *len_ind);
+	virtual int GetLobPart(long row, long column, void *data, long length, long *len_ind);
 	
 	// Get the list of available tables
 	virtual int GetAvailableTables(std::string &select, std::string &exclude, std::list<std::string> &tables);
